+++ source/engine/engine/CMakeLists.txt
@@ -1,5 +1,5 @@
 target_sources(${NC_ENGINE_LIB}
     PRIVATE
-        ComponentRegistration.cpp
         NcEngineImpl.cpp
+        RegistryFactories.cpp
 )
