--- source/engine/engine/ComponentRegistration.h
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace nc
-{
-class Registry;
-
-// Register all engine-side Component types with the Registry.
-void RegisterEngineComponents(Registry& registry);
-} // namespace nc
